@@ -1,17 +1,17 @@
 name: flutter_chat_list_view
 description: ChatListView is a listview designed for IM usage. 
-version: 0.0.1
+version: 0.0.8
 author: JimmyTai
 homepage: https://github.com/JimmyTai
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  collection: ^1.14.5
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   pedantic: ^1.8.0
